submodules(install): Add install script
Script that install the different submodule present.
This allows me to easily install/update the submodules without having to reread the instructions.
It works both for a fresh install and for a new version of the submodule.
The scripts puts lightweight git tag to see if it needs to reinstall or not.

I took the install procedure in each submodule 'README' and included them in the script.