bash: Replace 'll' alias by a function
The 'll' alias did not allow to pass in parameters.
The functions can take multiple parameters now,
so we can list a different directory from the current one.

Another thing to note, was that any file with whitespaces in the name would screw the 'column' formatting
This is fixed by using a separator other than whitespaces.
I made the separator be commas ','. My assumption being that it would not appear anywhere else.
I did not pick semicolons ';' because trying so, it would remove colors that ls outputs,
as ANSI escape code for colors contain semicolons.

In '.bashrc' I also moved the place where we set aliases for colors.
I put it before sourcing bash functions, for the same reason as other aliases.
Functions might depend on aliases, in this case, the 'll' function depends on the 'ls' alias