neovim(typing.vim): Plugin to ease typing of certains characters
It inserts a pair of: double quotes, single quotes, parentheses, braces and square brackets
But will return a <Right> keycode when trying to close the pair if the next character is already the one typed

This is to match my experience with Intellij, which would do something like this by default

A global function 'Get_char' was added, it returns the character in front of the cursor.
It was made global because of how general and useful it is.