config(dircolors, readline): Replace all use of color code '40' for '49'
'49' reset the background color to the 'default' background color of the terminal

After trying a transparent background in alacritty,
the background would sometime be fully colored.
This is due to the use of color code '40' which is the black color of the terminal
and is not affected by transparency in alacritty.

So before, the 'default' color and 'black' were the same.
But with transparency, they are not, and only 'default' is transparent