bash(ll): 'll' accepts '-a' option from 'ls (1)'
This enable the listing of hidden with the formatting of 'column (1)'

Because the function has to loop through each argument,
we can't simply pass the '-a' option with something like:
`ls -l $@`

This implementation parses the 1st argument to find if the option is present or not.
Then for each invocation of 'ls (1)', it passes the option if it was present.