git: Add 'gitconfig' at root of project
This config file is part of the repo so it is versioned and shared on all machines
To trick git into automatically using it as a configuration, I use the '-c' option in my alias

Git as a configuration property, named "include.path" to include another configuration
on top of the usual system, global and local files.
This property is set to the path of the file, i.e. "~/gitconfig".
And so git will read the extra configuration there.

The assumption is that the repo is checked out as a home folder.