config(bash): Add '--work-tree' option to 'dotfiles' alias
Whenever I was in a directory different from the root of the project,
git would display that all files were removed when running 'git status'

After reading 'git (1)' man page,
It is mentioned that for '--git-dir' option, the discovery of the project root is off.
This means that it assumes the work tree is the current dir, not where the git dir is at,
like I assumed

So the solution, as mentioned in the man page as well, is to use the '--work-tree' option.

The '-C <dir>' option won't work for my use case, since it does not use relative path of the current dir
but only that on the <dir> passed as argument,
so it's not useful as a global alias