config(python): Python PS1 prompt showing variables
This prompt shows global variables and their values.
It only shows "new" variables, that were introduced from the interactive session.
This is done by recording the variable in the global scope when this file is first sourced
and every time the PS1 prompt is refreshed, it compares the current global scope to the recorded one.
The diff is taken and pretty printed as the prompt

Disclaimer: I barely use Python, probably bad