bash(ll): Add '-d' option to ll, same as "ls (1)"
With '-d' we can list the permission of a directory only but not its content.

Also simplified the conditional. If the the argument is a directory but not a symlink to a directory
then we remove the first line, which is the total of files in the directory, like "total: 123"
For the rest of the cases, we are listing a single file, so no need to remove the first line