config(bash): Add default PS1 prompt when program 'prompt' is not found
My PS1 requires a program named 'prompt' to be on the $PATH
At this current time, the program is not in this repository
therefore, cloning this repository and using its bashrc and other env var
will result in a broken PS1 prompt.

Now, are default simple prompt will be used it not found.
We have to set '$PROMPT_COMMAND' to the empty string,
because the command it points to will set the PS1 prompt to the broken one on each command executed