gitmodules: Change git url from ssh transport to https
On new installs, you can not fully clone this repository and its submodules.
With an ssh url, you need to have ssh keys set up to just clone the repository.
Which is not the case on a new device.

Whereas, with https, you don't need to be authenticated or a special permisson to clone a public repo from github
This makes it easier for a fresh install to clone this repo and its submodules