submodules: Add missing dependencies in script
There were missing dependencies, and the installation did not work when testing inside a VM
For 'dwm' and 'dmenu', theres are dependencies to some X11 libs,
and added 'asciidoc' because 'git' needs it to create documentation.

Also fix a little bug when creating the final string of packages to install

  Each submodule has a string of the different packages it needs for installation,
where each package is just separated by white space.
A final string of all the packages needed to install multiple submodules,
is made by concatenating the different strings, which is passed to 'apt install'.

  The problem was when concatenating the string, there was no space between two different strings.
So the last package from the 1st module and the first package from the 2nd module would be concatenated into a single word
This is fixed by adding a space after each string of packages when building the final string