neovim(commands): Refactor commands and overall structure
Put the mappings <leader>cc and <leader>cd in commands.vim
With my custom commands, there is usually but not always an accompanying map
that makes its easier to use the command.
Such is the case for '<leader>cc' and '<leader>cd'
and having them next to the definition of the command they call, makes it
easier to modify both the mapping and the command

Add the -nargs option when creating my custom command.
It is mostly for fluff, I dont personally need it too much, but it is good practice.

Add a new command Lcd
It changes the directory for the window only, to the folder containing the current file/buffer
Same as the "Cd" custom command but only local to the window