Add alias and env var for wget and gnupg
'Wget' and 'GNUPG' add a file and folder respectively in the home folder.
In a continuous effort to clean my home folder and adhere to the XDG convention,
I removed the file and folder and followed the Arch Linux wiki on how to prevent programs
from poluting the home folder.