neovim(map): new mapping for quickfix/location list and buffers
* Add mappings to navigate the quickfix list, the location list and the buffer list
Those two list are very powerful features of vim and I want to use them more.
And want to be able to quickly navigate through a long list of buffers.

* Delete ftplugin for help files.
The function keys were used there and conflicting with my quickfix/location list mappping.
The same behavior of the previous mapping can be achieved with the location list.
When using 'g0' in a help file, by default, it creates a location list for the chapters.

* Add mapping to go forward the tag list
It serves as the opposite of Control-T which goes backward the tag list,
to the beginning before jumping to tag to tag to tag.

* Add mapping starting with comma to close buffer and quickfix list
The user manual recommend using ',' as a prefix for mapping since it is rarely used.
This is a test to see if it is more convenient to use it,
since :bdel and :cclose are short commands already.