bash(ll): Refactor ll function and simplified it
I didn't like the look of my old 'll' function. It looked like a smell.

I extracted the part that does the formatting and prettying into its own function
that can then be used by an alias or another function that itself will call 'ls'
This is more modular

For now, I have a simple 'll' function that calls ls -l
and simply pass the output to the prettier, formatting function.