bash(alias): Cloned 'alias_completion' and use it for 'df'
Bash aliases can't access the completion of the underlying command.
i.e. if we were to type the actual command and not the alias, we would have completion
but with the alias, not.

@CykerWay made a script that allows to access the correct completion for aliases,
called "alias_completion"

I changed the order of some bash script sourced in '.bashrc'.
"alias_completion" requires some helpers function from the bash_completion package
And so we source "bash_completion" first.