nvim(typing): Refactor how vim ft dont match double quotes
The autocmd to unmap the " mapping was broken.
If we started in a vim file, the mapping was never created and we got an error
We also got an error if we already remove the mapping, left and rejoined another vim buffer

For now, I just added logic to the underlying function
so that for vim ft, double quotes wont be matched

If there are more filetypes that need that change, another technique will be used.