(submodule) Add alacritty submodule
New program installed from source, alacritty.
It offers documentation and better support if installed from source.
Install script is updated with instruction to set up alacritty and its dependencies.
This version came with a bash-completion script which is now sourced in '.bashrc'
And also, this new version of alacritty replaced .yaml with .toml, so my config had to change.