(alacritty) Add a local config file to .gitignore
This file is mainly use to set the font size.
The different machines I use have different screen sizes
and I use a different font size on each.

It also sets the opacity for alacritty. Because I frequently change the value
and don't want git to complain about a change in values.
I updated my bash function to update the opacity for the new file