(gitignore) Invert the ignore logic for the config folder
Instead of ignoring everything in .config and whitelisting the folder I want to track
I decided to invert my .gitignore for that folder.
So any new folder in .config, git will mention as untracked file

I had a problem forgetting to 'git add' some new config since git fully ignored it.
I would have to first opt-out from ignoring it and then add it to the repo.
With this new system, my configuration files will be more in sync