(ll) Add -h flag to for nicer file size
The '-h' flag display file size in human readable form.
It uses common unit, like K, M, G and a small number in front.
That number can be decimal with a comma.

That comma messes with the 'pretty_ll' function which uses
commas as a delimiter between each column of text.
I changed the delimiter to be a backslash '\' which is a '\\' for 'awk' and 'column'.
This wont tabulate nicely if a file contains a backslash. But have not found a better solution yet.