(build_run_test) Windows are thinner and disabled local options
Made the output windows smaller, to take only 33% of the width of its original window.
Also set some window local options for them, when the first window get created.
I removed number and the fold column, to make more place for the actual text.
Also removed cursor highlighting.
And set wrap to true, because I don't want to switch window and move the cursor
to scroll the text left/right