MAJOR: Rename gitdir and move some root folder
The main gitdir 'dotfiles' has been renamed to the conventional '.git'
and a new folder 'dotfiles' was created to contain most of the repository root folders,
like 'tools/', 'submodules/', and 'img/' but also the gitconfig file

This was needed as I was creating more and more folder at the root.
It did not feel right to make a main folder in like '.local/share'
I wanted it to be at the root, and 'dotfiles' felt like the right name
but it was already taken by the gitdir.
It was a mistake to create a gitdir with another name and be fancy with a bash alias.
Being stupid simple was the right approach.

This will also for a more structured repository where files/folders that don't have
a place in the traditional home folder, with XDG sub folders, can find their place
in the 'dotfiles' folder. For example, any file needed to be installed or symlink outside the HOME folder.

In terms of migrating the repo. The submodules have been deinit and reinit.
This is because each submodule '.git' file contains a path to the root repository gitdir
i.e. it referenced the previous gitdir name 'dotfiles' and it has to change to '.git'
Any mention to the old path, for 'tools', 'submodules', 'img' or 'gitconfig' have been updated.
So mainly environment variables, scripts and bash aliases