(tmux) Make tmux use nvim instead of neovide
Made tmux shells non-login and overrides env var
EDITOR, VISUAL and SUDO_EDITOR for the tmux server.

My workflow with neovide does not work inside tmux.
When I start neovide, I get a blocked terminal and
a new window popup with neovide. It is not swallowed by dwm
This is probably because tmux is not blocked when one of its
subshell starts a blocking gui.

The solution is for tmux to overrides EDITOR, VISUAL and SUDO_EDITOR
to nvim instead of neovide.
For this I have to change tmux to non-login, since those are set
in '.bash_profile' which is sourced for every login shell.