(alacritty,bash) Make "Paste" a bash keybind, not alacritty
It was a bad idea to give alacritty a keybind with Control-v
Because it prevents Neovim from doing visual-block.

I do want to have Control-v for paste in the terminal and that is
best implemented by Bash as a readline function.
Because it won't interfere with Neovim.

Neovide can keep Control-v as Paste in insert mode for the moment.