(bash) Convert bash function to external scripts
I want to rewrite my scripts in other languages and will need to have them as external programs.

Removed exported functions, since I don't need them to be exported and make the output of 'env' harder to read.

Replaced some completion functions with completion word list.
The list of word is static, so the function is basically run once to generate the word list which is used by 'completion'
This avoids running the function each time completion is attempted