(fzf) fzf plugin for bash/neovim
Fuzzy finding script for finding files/editing files/changing directory/opening buffers

There is a main external script that starts fzf in a tmux popup window or another terminal
And then bash and neovim use that external script to do the fuzzy finding.
Then they simply use the result to modify their internal state.

For neovim, I prefer this over Telescope. I can get the same functionalities from fzf
and it is consistent with bash.

There are two git alias, for 'ls-files' and 'ls-tree', simply so the output of those commands
are rooted at the base of repository. Which allows to create absolute paths for bash and neovim.

Add a fzf.toml alacritty config file, to format the additional term window