(neovide) Add plugin to switch 'guifont' options
This is very similar to my 'fonts' script. But this time for neovide
My 'fonts' script changes the font for alacritty which does not affect neovide
So another plugin had to be written.

I could have a common program that both alacritty's font and neovide's font would both
use and that way it would be synced. But instead, having it this way, I can have a
different font in the terminal and in Neovide. Also, I can now use a Hydra to switch fonts.