(git) Fix 'includeIf gitdir' conditional
The 'includeIf' directive was not working as intended. Subdirectories with a gitdir
not inside the HOME folder would still have the optional gitconfig included.
For example '~/foo/bar/.git' would include the gitconfig

After rereading the man page, git-config(1), a 'gitdir' conditional that ends
with '/' works by looking recursively all the subdirectories given. So '~/foo/bar/.git' is
a subdirectory of '~/'. It works exactly the same as if '**' was appended.

My original intent was to only include the gitconfig if the gitdir is right inside the
HOME folder, something like '~/some-git-folder'. This is in case, the git folder is named differently
via the '--git-dir' option. The gitdir conditional takes glob patterns.