Update neovim submodule since it was forced-pushed
I forced-pushed the neovim submodule because I rewrote the commit message
I want them to explain the step-by-step taken to create it.

It originally had simple commit messages because the repository was private
and not imported by this one.
I forgot to rewrite the commits when importing it and making it public.