Replace '.config/nvim' by submodule with config and packages
I separated my neovim config into a new repository:
https://github.com/RyanBeerBarron/neovim-config
And archived my previous neovim repository, which contained only the packages:
https://github.com/RyanBeerBarron/neovim-packages

This new repository contains all of my neovim configuration which was contained in this repository.
And the history thanks to 'git filter-repo'. As well as the all the neovim packages I use,
still as submodule inside the 'pack/' folder.

How it was done:
This repository removed the old, archived submodule by doing 'git submodule deinit' and removing the entry from '.gitmodules'
And removing the '.config/nvim' folder with 'git rm -r'. This has the benefit of keeping the git history.

For creating the new repository, neovim-config, it started as a clone of this repository and with 'git filter-repo',
I removed everything besides the neovim config, rooted the repository in the 'nvim' directory and integrated the old submodule inside it.
This was done by deinitializing all the submodule. Modifying the .gitmodules file and reinitializing everything.