Change XDG CONFIG_HOME, CACHE_HOME and '.local' to visible directories
"./.config moved to ./config"
"./.cache moved to ./cache"
"./.local moved to ./local"

I prefer having those directories visible. In general hidden files/directories should be avoided
or only used if really meant to be hidden to the end user.
But for the config folder and cache, it is important to know about them.
I use the config and local folder daily, to edit my config files and such. So having it hidden does not make sense to me
since they are the directories I spent the most time within

Same reasoning as the previous commit. In my mind, it should not be hidden