(workspace) Rename script 'projects' with 'workspace'
The name change was to be closer to 'workspaces' within the LSP specs and how 'nvim' sets up language servers.

Simplified script in the process, now there is only one line of output instead of two.
The removed line was only used for workspaces using git worktrees.
It was not necessary to output it and only made usage of the script more complicated.
As I had to split the output lines.

Add a check on shell startup (.bashrc), to verify shell is in a workspace,
before setting up the configuration around workspaces (creating directories, adding keybinds, etc...)