(tmux, nvim) Enable one nvim instance per tmux session/workspace
This adds a wrapper script for 'nvim', named 'nvim'
A helper script for future wrapper script named 'next-exec'
And update the neovim-config submodule to a new commit which allows to disable the workspace plugin

The wrapper script helps with having only one instance of nvim per workspace and
makes tmux automatically focus that instance whenever 'nvim' is invoked in another window/pane.

This is achieved by using the neovim server features with '--listen' and '--remote'.
It solves my problem of having multiple nvim instance in the same workspace opening up the same buffers
and getting an error messages each time I open nvim. Now only one nvim instance per workspace will exists
and tmux will automatically focus the nvim instance.
The unix socket created for the server has the tmux pane id encoded in its name.

The wrapper script makes use of the new commits in my neovim-config submodule, to disable workspace plugin when editing
git commits.

'next-exec', the helper script, is meant to be used any future wrapper script.
Given the path of a wrapper script ($0 is set to that value automatically) it will return the next absolute path to
a program of the same name that comes AFTER in the PATH environment variable.