(tmux) Disable 'monitor-bell' and 'monitor-activity'
'monitor-bell' was actived by default and would add a '!' on the status line
whenever a program would generate a bell in a non-focused pane/window.
Which I don't really care for and happened too often for me to care.

'monitor-activity' was by default turned off, so it was added to the configuration for exhaustiveness