(bash) Configure bash 'history'
- Set history envvars to configure the history system. Most notably:
    * 'HISTIGNORE' to not log certain commands to the session history or hisory file.
            Ignored commands would any command that is always the same and very often entered
    * 'HISTFORMAT' to add a formatted timestamp to each history entry and using the 'history' builtin.
- Use 'history -a' to automatically new commands to the histfile (done inside 'PROMPT_COMMAND')
        This allows any new bash process to directly inherits the history of another bash
        process that has not ended yet. Also the values are appended and don't overwrite the history file
- Set the shopt 'histappend' to append the history of the current bash process instead of overwriting the histfile

I verified the history appending mechanism and it does not add the same log multiple time in the history file. As one would hope.
If a log is already present, history does not add it again. The history file keeps timestamps of when each command happened.
And I suppose it uses that timestamp to verify if a command was already entered.