(workspace) Revamp 'workspace' script
Did not want my home directory to be a workspace so revamped the script to have more control over what is and is not a workspace.
It does not rely on markers (like a '.git' directory) anymore to determine where a workspace starts.
Instead, any directory can be a workspace and the users can add a workspace or list them all

This is commit also updates my nvim config. It changed the functioning of its 'workspace' lua plugin.
see https://github.com/ryanbeerbarron/neovim-config/commit/54a6eab (Rewrite 'workspace' plugin completely, 2024-07-21).

For 'nvim' to start in a workspace, it needs two conditions:
- a) But inside a workspace
- b) Set the global variable 'long_session' to a true value. Setting this variable is done by the helper script 'nvim_session'

This setup allows to create a nvim instance for work, (enables the 'workspace' plugin, starts LSP, uses custom config)
but also a default, lightweight, nvim instance for quick edits. Even if started from a workspace