(java) Replace '_JAVA_OPTIONS' envvar with alias
The presence of the envvar causes every 'java' invocation to echo 'Picked up _JAVA_OPTIONS' on stderr.
I wanted to remove this annoying message.

With the VM option '-Djava.util.prefs.userRoot', we get the same behavior.
I added an alias that sets the VM option, and removed the envvar which causes the message.