(nvim) Revert the overload script 'nvim'
The idea is there but the current implementation is too annoying.
Most of the time I have one long running 'nvim' instance up,
and the subsequent started 'nvim' process are meant for a quick edit.

Having the quick edit done in the main 'nvim' instance is confusing.
- I lose track of which buffer was open my main 'nvim' instance.
- I lose track of the tmux pane I was before the script switched automatically.
- If I quickly do my edit, and type ":wq", I close my main 'nvim' instance and need to reopen and wait for LSP startup.

The script has some neat ideas for tmux so I'm keeping it around and we'll come back to it.